requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
httpx==0.27.2
python-dotenv==1.0.1
requests==2.32.5
google-auth==2.34.0
email-validator==2.3.0